+++ R/RcppExports.R
@@ -213,8 +213,8 @@ cpp_hash_string <- function(x) {
     .Call(`_fixest_cpp_hash_string`, x)
 }
 
-cpp_md_markup <- function(Rstr) {
-    .Call(`_fixest_cpp_md_markup`, Rstr)
+cpp_escape_markup <- function(Rstr) {
+    .Call(`_fixest_cpp_escape_markup`, Rstr)
 }
 
 cpp_get_nb_threads <- function() {